Adding many tests to cover the changes to the cgroup logic. Cgroup logic changed so it uses the systemtools' cgroup directory which will point to /host/cgroup. This is important as I found many jira tickets with issues finding the cgroup stats files but I found that we can reliably use /host/cgroup when it is set to mount the host's /sys/fs/cgroup directory.
Otherwise, I also changed some of the logic so it will check for the
subsystem in the cgroup file rather than the previous method of just
using the first one. This should allow me to add detect and add a
fallback for these scenarios.

The issue with the previous fallback code was that it assume
/proc/<pid>/root/sys/fs/cgroup was the standardized way but I cannot
find any documentation that this is standard. The only standard is
/sys/fs/cgroup is reliably mounted by distributions using systemd.
See
https://github.com/systemd/systemd/blob/main/docs/CGROUP_DELEGATION.md
for documentation on this path.

So hence the motivation to move to using /host/cgroup when we can
instead of the container's /sys/fs/cgroup which may or not be properly
mounted for us. This seems to depend on container runtime whereas the
host's /sys/fs/cgroup is well understood and containers runtimes are
able to delegate with this spec.