NVSHAS 7990 7925 8181 - cgroup error fixes
The changes will try to minize the amount of logging when the stats file
getCgroupPathReaderV2 no longer returns a default path, it returns ""
instead.
cgroup_test modified to deal with the change.
fillContainerProperties() will test the path at start up to check if it
exist. If it doesn't work, then we force the path to be "" so we don't
try to get stats.
updateContainerStats() checks for a non empty cgroup path. It also does
error handling and messaging here.
getMemoryStats() returns an error on file not found so we no longer
silence errors.