NVSHA-8283 - Dropping log levels for RunWithTimout() calls
We had an issue where the pathwalker timeout will flood the logs, making
it difficult to use. I dropped the log level to debug because we need it
for debugging but we don't need it for normal operation. The timeouts
are informational but do not actually mean there is an error. The
pathwalker operation will eventually restart so having the logs about
starting/timeout, etc will just flood the logs but not provide useful
insight in any problems.